Added backgroundUnicorn
If someone can give me the commands / lines for one script to have root (Backgroundunicorn) so it can talk with the unicornhat. And also allow me to call/comunicate with finalpiprogram which has normal pi permisions. also backgroundunicorn has to run on startup. I have lost the ability to care for this.